+++ packages/components/src/AppLayout.tsx
@@ -1,7 +1,7 @@
 /** @format */
 'use client';
 
-import { CookieProvider, EnvironmentProvider, PlatformProvider, SidebarProvider, SupportProvider, ThemeProvider, ToastProvider, ShortcutsProvider } from '@xernerx/providers';
+import { CookieProvider, EnvironmentProvider, PlatformProvider, ShortcutsProvider, SidebarProvider, SupportProvider, ThemeProvider, ToastProvider, UserProvider } from '@xernerx/providers';
 
 import { CookiePrompt } from './CookiePrompt';
 import { Page } from './Page';
@@ -9,10 +9,10 @@ import React from 'react';
 
 export function AppLayout({ children }: { children: React.ReactNode }) {
 	return (
-		<>
-			<EnvironmentProvider>
-				<PlatformProvider>
-					<ThemeProvider>
+		<EnvironmentProvider>
+			<PlatformProvider>
+				<ThemeProvider>
+					<UserProvider>
 						<ShortcutsProvider>
 							<ToastProvider>
 								<CookieProvider>
@@ -26,9 +26,9 @@ export function AppLayout({ children }: { children: React.ReactNode }) {
 								</CookieProvider>
 							</ToastProvider>
 						</ShortcutsProvider>
-					</ThemeProvider>
-				</PlatformProvider>
-			</EnvironmentProvider>
-		</>
+					</UserProvider>
+				</ThemeProvider>
+			</PlatformProvider>
+		</EnvironmentProvider>
 	);
 }