+++ packages/providers/index.ts
@@ -9,3 +9,4 @@ export * from './src/ToastProvider';
 export * from './src/SupportProvider';
 export * from './src/EnvironmentProvider';
 export * from './src/ShortcutsProvider';
+export * from './src/UserProvider';