+++ packages/ui/index.ts
@@ -2,3 +2,4 @@
 
 export * from './src/Button';
 export * from './src/Toggle';
+export * from './src/Divider';