+++ packages/ui/src/Divider.tsx
@@ -1,0 +1,8 @@
+/** @format */
+'use client';
+
+export function Divider() {
+	return (
+		<div className='relative mb-2 h-[1px] w-full before:absolute before:inset-x-0 before:bottom-0 before:h-[1px] before:bg-gradient-to-r before:from-transparent before:via-(--accent) before:to-transparent' />
+	);
+}