+++ apps/www/CHANGELOG.md
@@ -1,5 +1,15 @@
 # www
 
+## 0.2.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @xernerx/components@0.1.0
+  - @xernerx/providers@0.1.0
+  - @xernerx/lib@0.0.2
+  - @xernerx/feedback@0.0.2
+
 ## 0.2.28
 
 ### Patch Changes