+++ packages/providers/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @xernerx/providers
 
+## 0.1.0
+
+### Minor Changes
+
+- <!-- @format -->
+
+  Updated UI and added User Context
+
+  ## @xernerx/components
+  - Added UserProvider in AppLayout
+  - Fixed an issue where CookiePrompt was using the wrong accent color origin
+  - Updated Sidebar UI and seperated modals from sidebar
+
+  ## @xernerx/providers
+  - Updated ThemeProvider to inject current set theme before page load
+  - Fixed an issue where ToastProvider was using thw wrong accent color origin
+  - Added UserProvider for User Context
+
+  ## @xernerx/styles
+  - Added new accent specific colors
+
+  ## auth
+  - Fixed issues across multiple pages where the accent color was derived from the wrong origin
+
+  ## @xernerx/lib
+  - Fixed an issue where discord login credentials were not being refreshed.
+
+  ## @xernerx/ui
+  - Fixed an issue where the toggle switch had a misaligned toggle
+
+  ## bot
+  - Changed backend code
+
+### Patch Changes
+
+- Updated dependencies
+  - @xernerx/lib@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes